# api-gateway/docs/swagger.yaml
definitions:
  booking.BookingRequest:
    properties:
      checkInDate:
        type: string
      checkOutDate:
        type: string
      hotelid:
        type: string
      roomId:
        type: string
      roomtype:
        type: string
      totalAmount:
        type: integer
      userid:
        type: integer
    type: object
  booking.BookingResponse:
    properties:
      bookingId:
        type: string
      checkInDate:
        type: string
      checkOutDate:
        type: string
      hotelId:
        type: string
      roomId:
        type: string
      roomtype:
        type: string
      status:
        type: string
      totalAmount:
        type: integer
      userId:
        type: integer
    type: object
  booking.CreateWaitingList:
    properties:
      check_in_date:
        type: string
      check_out_date:
        type: string
      hotel_id:
        type: string
      room_type:
        type: string
      user_email:
        type: string
      user_id:
        type: integer
    type: object
  booking.DeleteResponse:
    properties:
      bookingId:
        type: string
      message:
        type: string
    type: object
  booking.GetUsersResponse:
    properties:
      bookingId:
        type: string
      checkInDate:
        type: string
      checkOutDate:
        type: string
      hotelId:
        type: string
      roomId:
        type: string
      roomType:
        type: string
      status:
        type: string
      totalAmount:
        type: integer
    type: object
  booking.GetWaitingResponse:
    properties:
      check_in_date:
        type: string
      check_out_date:
        type: string
      hotel_id:
        type: string
      id:
        type: string
      room_type:
        type: string
      status:
        type: string
      user_email:
        type: string
      user_id:
        type: integer
    type: object
  booking.UpdateRequest:
    properties:
      bookingId:
        type: string
      checkInDate:
        type: string
      checkOutDate:
        type: string
      roomId:
        type: string
      roomtype:
        type: string
      status:
        type: string
    type: object
  booking.UpdateWaitingListRequest:
    properties:
      check_in_date:
        type: string
      check_out_date:
        type: string
      hotel_id:
        type: string
      id:
        type: string
      room_type:
        type: string
      user_id:
        type: integer
    type: object
  booking.WaitingResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server for a restaurant reservation system.
  title: Hotel Booking System
  version: "1.0"
paths:
  /bookings:
    post:
      consumes:
      - application/json
      description: Create a new Booking
      parameters:
      - description: Booking request body
        in: body
        name: hotel
        required: true
        schema:
          $ref: '#/definitions/booking.BookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create a new booking
      tags:
      - Booking
  /bookings/{id}:
    delete:
      description: Remove a booking from the system
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.DeleteResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete a booking
      tags:
      - Booking
    get:
      description: Retrieve a specific booking by its ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get booking by ID
      tags:
      - Booking
    put:
      consumes:
      - application/json
      description: Update the details of an existing booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/booking.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update a booking
      tags:
      - Booking
...
  /users/{id}/bookings:
    get:
      description: Retrieve all bookings made by a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.GetUsersResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get bookings for a user
      tags:
      - Booking
...
  /waitinglist:
    post:
      consumes:
      - application/json
      description: Add a new booking to the waiting list
      parameters:
      - description: Waiting list request body
        in: body
        name: waitinglist
        required: true
        schema:
          $ref: '#/definitions/booking.CreateWaitingList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.WaitingResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Add a booking to the waiting list
      tags:
      - WaitingList
  /waitinglist/{id}:
    delete:
      description: Remove an entry from the waiting list
      parameters:
      - description: Waiting list ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.WaitingResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete a waiting list entry
      tags:
      - WaitingList
    get:
      description: Retrieve a specific entry from the waiting list by its ID
      parameters:
      - description: Waiting list ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.GetWaitingResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get waiting list by ID
      tags:
      - WaitingList
    put:
      consumes:
      - application/json
      description: Update the details of an entry in the waiting list
      parameters:
      - description: Waiting list ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated waiting list details
        in: body
        name: waitinglist
        required: true
        schema:
          $ref: '#/definitions/booking.UpdateWaitingListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.WaitingResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update a waiting list entry
      tags:
      - WaitingList
securityDefinitions:
  Bearer:
    description: Enter the token in the format `Bearer {token}`
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"